+++ src/features/room/index.ts
@@ -1,2 +1,3 @@
+export { default } from "./index.vue";
+export { Register } from "./ui";
 export { default as useRoomStore } from "./store/useRoomStore";
-export * from "./ui";
