+++ src/features/room/ui/index.ts
@@ -1,3 +1,4 @@
+export { default as Layout } from "./layout.vue";
 export { default as RoomList } from "./room-list.vue";
 export { default as Register } from "./register.vue";
 export { default as MakeRoom } from "./make-room.vue";
