+++ src/pages/index.vue
@@ -5,25 +5,18 @@
 
   <div class="q-pa-md row justify-center">
     <layout>
-      <template #header> <search /> <Title /> </template>
+      <template #room>
+        <room />
+      </template>
 
-      <template #side-header> <make-room /> </template>
-      <template #side> <room-list /> </template>
-
-      <template #connect>
-        <connect
+      <template #chat>
+        <chat
           v-if="!!selected_room?.id"
           :connecting="connecting"
           :room="selected_room"
           :current_user="current_user"
         />
       </template>
-
-      <template #content> <content /> </template>
-
-      <template #toolbar> <actions /> </template>
-
-      <template #footer> <submit /> </template>
     </layout>
   </div>
 </template>
@@ -31,8 +24,8 @@
 <script setup>
 import { storeToRefs } from "pinia";
 
-import { Register, MakeRoom, RoomList, useRoomStore } from "@/features/room";
-import { Connect, Content, Submit, Actions, Title, Search } from "@/features/chat";
+import Room, { Register, useRoomStore } from "@/features/room";
+import Chat from "@/features/chat";
 import Layout from "./ui/layout.vue";
 
 // feature간 데이터 공유
